Further organize create completion
