+++ GPTifier/include/commands.hpp
@@ -2,4 +2,3 @@
 
 void print_help_messages();
 void print_build_information();
-void command_run();
